Fix CI file, __init__.py, imports issues, tests and small stuff in .gitignore and setups.py
