+++ .gitignore
@@ -1,4 +1,3 @@
-.DS_Store
-.vscode
+
 *.pyc
 libsvmdata.egg-info/*
